fix: updated spot balances holding.
* refactor: lifted exchange tokens query up to Overview component.
* fix: filtered same synth tokens from 1inch tokens.
* fix: token description.
